+++ src/main/java/frc/robot/subsystems/led/Leds.java
@@ -1,5 +1,8 @@
 package frc.robot.subsystems.led;
 
+import java.time.Duration;
+import java.time.LocalTime;
+
 import edu.wpi.first.wpilibj.AddressableLED;
 import edu.wpi.first.wpilibj.AddressableLEDBuffer;
 import edu.wpi.first.wpilibj.util.Color;
@@ -196,6 +199,37 @@ public class Leds extends SubsystemBase {
     }
 
     // ========================= Function Commands =============================
+
+    /**
+     * Blink the LEDs to specified Color and then return to Alliance color
+     * 
+     * @param color color being blinked to
+     * @return Blink Command
+     */
+    public Command blinkCommand(Color color) {
+        Duration WAIT_TIME = Duration.ofMillis(500);
+        Command blinkCommand = new Command() {
+            LocalTime startTime;
+
+            @Override
+            public void initialize() {
+                setColor(color);
+                startTime = LocalTime.now();
+            }
+
+            @Override
+            public boolean isFinished() {
+                Duration timeWaited = Duration.between(startTime, LocalTime.now());
+                return timeWaited.compareTo(WAIT_TIME) >= 0;
+            }
+        };
+        blinkCommand.addRequirements(this);
+        return CommandUtils.addName(blinkCommand);
+    }
+
+    public Command setAllianceColorCommand() {
+        return new InstantCommand(this::setAllianceColor);
+    }
 
     /**
      * Set the lights to a scrolling pattern
