--- src/main/java/frc/robot/util/KColor.java
@@ -1,8 +1,0 @@
-package frc.robot.util;
-
-import edu.wpi.first.wpilibj.util.Color;
-
-public class KColor extends Color {
-    public static final Color ALLIANCE_BLUE = new Color(0, 0, 255);
-    public static final Color ALLIANCE_RED = new Color(255, 0, 0);
-}
